GeoloniaMapコンポーネントのリファクタリングとマーカー管理機能の追加
